requirements.txt
langchain 
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
fastapi
uvicorn
python-dotenv
pypdf2 
langchain-google-genai
langchain-huggingface
sentence-transformers
bcrypt
deep-translator
twilio
requests